test: consolidate n-api test addons
It takes time to build each of the addons used to test n-api.
Consolidate a few of the smaller ones to save build time.
