+++ test/addons-napi/test_general/testGlobals.js
@@ -2,7 +2,7 @@
 const common = require('../../common');
 const assert = require('assert');
 
-const test_globals = require(`./build/${common.buildType}/test_globals`);
+const test_globals = require(`./build/${common.buildType}/test_general`);
 
 assert.strictEqual(test_globals.getUndefined(), undefined);
 assert.strictEqual(test_globals.getNull(), null);
