+++ test/addons-napi/test_general/test_general.c
@@ -33,11 +33,40 @@ napi_value testGetVersion(napi_env env, napi_callback_info info) {
   return result;
 }
 
+napi_value doInstanceOf(napi_env env, napi_callback_info info) {
+  size_t argc = 2;
+  napi_value args[2];
+  NAPI_CALL(env, napi_get_cb_info(env, info, &argc, args, NULL, NULL));
+
+  bool instanceof;
+  NAPI_CALL(env, napi_instanceof(env, args[0], args[1], &instanceof));
+
+  napi_value result;
+  NAPI_CALL(env, napi_get_boolean(env, instanceof, &result));
+
+  return result;
+}
+
+napi_value getNull(napi_env env, napi_callback_info info) {
+  napi_value result;
+  NAPI_CALL(env, napi_get_null(env, &result));
+  return result;
+}
+
+napi_value getUndefined(napi_env env, napi_callback_info info) {
+  napi_value result;
+  NAPI_CALL(env, napi_get_undefined(env, &result));
+  return result;
+}
+
 void Init(napi_env env, napi_value exports, napi_value module, void* priv) {
   napi_property_descriptor descriptors[] = {
     DECLARE_NAPI_PROPERTY("testStrictEquals", testStrictEquals),
     DECLARE_NAPI_PROPERTY("testGetPrototype", testGetPrototype),
     DECLARE_NAPI_PROPERTY("testGetVersion", testGetVersion),
+    DECLARE_NAPI_PROPERTY("doInstanceOf", doInstanceOf),
+    DECLARE_NAPI_PROPERTY("getUndefined", getUndefined),
+    DECLARE_NAPI_PROPERTY("getNull", getNull),
   };
 
   NAPI_CALL_RETURN_VOID(env, napi_define_properties(
