--- test/addons-napi/test_globals/binding.gyp
@@ -1,8 +1,0 @@
-{
-  "targets": [
-    {
-      "target_name": "test_globals",
-      "sources": [ "test_globals.c" ]
-    }
-  ]
-}
