--- test/addons-napi/test_instanceof/binding.gyp
@@ -1,8 +1,0 @@
-{
-  "targets": [
-    {
-      "target_name": "test_instanceof",
-      "sources": [ "test_instanceof.c" ]
-    }
-  ]
-}
