--- test/addons-napi/test_instanceof/test_instanceof.c
@@ -1,28 +1,0 @@
-#include <node_api.h>
-#include <stdio.h>
-#include "../common.h"
-
-napi_value doInstanceOf(napi_env env, napi_callback_info info) {
-  size_t argc = 2;
-  napi_value args[2];
-  NAPI_CALL(env, napi_get_cb_info(env, info, &argc, args, NULL, NULL));
-
-  bool instanceof;
-  NAPI_CALL(env, napi_instanceof(env, args[0], args[1], &instanceof));
-
-  napi_value result;
-  NAPI_CALL(env, napi_get_boolean(env, instanceof, &result));
-
-  return result;
-}
-
-void Init(napi_env env, napi_value exports, napi_value module, void* priv) {
-  napi_property_descriptor descriptors[] = {
-    DECLARE_NAPI_PROPERTY("doInstanceOf", doInstanceOf),
-  };
-
-  NAPI_CALL_RETURN_VOID(env, napi_define_properties(
-    env, exports, sizeof(descriptors) / sizeof(*descriptors), descriptors));
-}
-
-NAPI_MODULE(addon, Init)
